elf: Monte-Carlo ASLR algorithm
There is a need to implement ASLR to make control-hijacking attacks difficult.
Unfortunately previous ASLR implementation is poor, it does not provide
randomized gaps between dynamic libraries.

When mapping shared ELF binaries (ET_DYN), mmap() function is used.
Its first argument, mappref/addr, when it is not NULL, is taken as a hint by
the POSIX-compliant kernel about where to place the mapping.

Use this feature to implement an opportunistic non-invasive ASLR.

1) Assign a reasonable range of addresses below the base of ld.so
to ASLR space

2) Randomly generate 'mappref' inside this ASLR space

3) Pass generated random 'mappref' as mmap() first argument
when mapping the memory for ELF image

If kernel's mmap() implementation encounters a collision with previously
mapped memory range, it will assign a suitable address range just like
it happens when NULL is passed as a first argument of mmap().

This is a Monte-Carlo randomized algorithm, it shows good results despite
its simpicity.

This feature can be turned on with the following environment variables:

LD_ASLR
      Set to non-empty value to turn ASLR on. ASLR space size is set to
      the default value (0x2000000000 for 64-bit, 0x8000000 for 32-bit)
LD_ASLR_SPACE
      Set to numeric value (e.g. 0x10000000) to turn ASLR on and set ASLR
      space size to this value

Signed-off-by: Eugene Korenevsky <eugene@ironmachine.tech>